PATCH THE translation between 'deviceType' and 'typeId'

+++ core/bus_center/lnn/net_builder/sync_info/src/lnn_exchange_device_info.c
@@ -58,7 +58,7 @@ static int32_t PackCommon(cJSON *json, const NodeInfo *info, SoftBusVersion vers
 static void UnPackCommon(const cJSON* json, NodeInfo *info, SoftBusVersion version)
 {
     char deviceType[DEVICE_TYPE_BUF_LEN] = {0};
-    uint8_t typeId;
+    uint16_t typeId;
     if (json == NULL || info == NULL) {
         SoftBusLog(SOFTBUS_LOG_LNN, SOFTBUS_LOG_ERROR, "para error!");
         return;
